# ScorecardAPI/openapi.yaml
openapi: 3.0.1
info:
  title: ScorecardAPI
  version: "1.0"
servers:
- url: /
paths:
...
  /api/Tournament:
    get:
      tags:
      - Tournament
      operationId: GetTournaments
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentOutputDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentOutputDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentOutputDTO'
    post:
      tags:
      - Tournament
      operationId: CreateTournament
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CreateTournamentDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTournamentDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateTournamentDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateTournamentDTO'
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CreateTournamentCompleteDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTournamentCompleteDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/CreateTournamentCompleteDTO'
  /api/Tournament/{tournamentId}:
    get:
      tags:
      - Tournament
      operationId: GetTournament
      parameters:
      - name: tournamentId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TournamentOutputDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentOutputDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/TournamentOutputDTO'
    delete:
      tags:
      - Tournament
      operationId: DeleteTournament
      parameters:
      - name: tournamentId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
...
components:
  schemas:
...
    CreateTournamentCompleteDTO:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int32
        matches:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/MatchOutputDTO'
      additionalProperties: false
    CreateTournamentDTO:
      type: object
      properties:
        generationMethod:
          type: integer
          format: int32
        pools:
          type: integer
          format: int32
        fighters:
          type: array
          nullable: true
          items:
            type: integer
            format: int32
      additionalProperties: false
...
    MatchOutputDTO:
      type: object
      properties:
        matchId:
          type: integer
          format: int32
        tournamentId:
          type: integer
          format: int32
        fighterOneId:
          type: integer
          format: int32
        fighterTwoId:
          type: integer
          format: int32
        fighterOneHealth:
          type: integer
          format: int32
        fighterTwoHealth:
          type: integer
          format: int32
        pool:
          type: integer
          format: int32
        duration:
          type: integer
          format: int32
        started:
          type: string
          nullable: true
        ended:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/MatchStatus'
      additionalProperties: false
    MatchStatus:
      type: integer
      format: int32
      enum:
      - 0
      - 1
      - 2
      - 3
...
    TournamentOutputDTO:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int32
        matches:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/MatchOutputDTO'
        fighters:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/UserOutputDTO'
        finished:
          type: boolean
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/UserResult'
          nullable: true
        disqualifiedFighters:
          type: array
          nullable: true
          items:
            type: integer
            format: int32
      additionalProperties: false
...
    UserOutputDTO:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        userId:
          type: integer
          format: int32
        tournaments:
          type: array
          nullable: true
          items:
            type: integer
            format: int32
      additionalProperties: false
    UserResult:
      type: object
      properties:
        userName:
          type: string
          nullable: true
        wins:
          type: integer
          format: int32
        losses:
          type: integer
          format: int32
        draws:
          type: integer
          format: int32
        sumRemainingHealth:
          type: integer
          format: int32
      additionalProperties: false